Rename home to student
