+++ entrypoints/student.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import {hydrate} from "react-dom";
 
-import StudentProfile from '../views/home.jsx'
+import StudentProfile from '../views/student.jsx'
 
 hydrate(<StudentProfile/>, document.getElementById("student_profile_body"));
